+++ src/steps/index.ts
@@ -3,6 +3,7 @@ import { userSteps } from './users';
 import { groupSteps } from './groups';
 import { deviceSteps } from './devices';
 import { applicationSteps } from './applications';
+import { ruleSteps } from './rules';
 
 const integrationSteps = [
   ...accountSteps,
@@ -10,6 +11,7 @@ const integrationSteps = [
   ...groupSteps,
   ...deviceSteps,
   ...applicationSteps,
+  ...ruleSteps,
 ];
 
 export { integrationSteps };